platform: nordic_nrf: Prevent transfers on uninitialized UART
Return an error if a transfer is attempted on a UART that has not been
initialized yet. Such attempt can occur when for instance an assertion
fails before the stdio UART is initialized. Since the nrfx_uarte driver
uses also an assertion for checking if it has been initialized prior to
starting a transfer, the CMSIS driver must prevent such attempts.
Otherwise, such situation would lead to a stack overflow and HardFault.

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>