platform: nrf5340: Add missing peripherals in spu_periph_init_cfg()
Add missing entries for SPLIT security peripherals (DPPIC and NVMC)
in the function providing the default configuration of peripheral
access permissions. Use NRF_REGULATORS and NRF_CLOCK instead of
NRF_OSCILLATORS and NRF_RESET (these peripherals have the same base
addresses, respectively) in that function, for consistency with what
is used for the nRF9160 platform. For the same reason, remove the
unnecessary entry for NRF_FPU.

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>