tfm core: make has_access_to_region() compatible with vendor IDAUs
We enhance the implementation of has_access_to_region(), so
it works with IDAU implementations, i.e. other than the ARM
SAU. We add a hook for a platform-specific check, in case
the memory address range crosses multiple IDAU boundaries.

We provide a declaration of the platform-specific hook in
SPM HAL API.

Change-Id: I18ad67637e561a54786846b844adde21240438f1
Signed-off-by: Ioannis Glaropoulos <Ioannis.Glaropoulos@nordicsemi.no>